Exclude Multi-Release files from Jackson while shading (#2793)
Cucumber-JVM uses Jackson Databind to write json. Because Jackson is a very
common dependency, to ensure we do not interfere with the system under test,
we shade it.

Unfortunately the Shade plugin occasionally gets updates and now includes
multi-releases. These were not included in the relocation pattern and
result in duplicate classes when Cucumber itself is shaded.

As `cucumber-core` is not a multi-release jar, we can exclude the multi-
release files altogether.

Fixes: #2786