[JUnit Platform] Fix UniqueId resolution
A unique id of any feature element is prefixed by the unique
id of the engine. This prefix, and only this prefix should be
skipped. Simply finding the first feature segment is not correct.

In practice, because we check if the unique id we try to resolve
is prefixed by the engine descriptor, everything does work out
fine. But this should help against malformed ids.