Fix copyright statement and use standard formatting
Aslak kept updating the year of the copyright notice, every year
until he removed it. Though it should be sufficient to provide
the first year of notice only.

Additionally, in 2013 he removed his personal copyright and assigned it
to the "Cucumber Organisation". Which is presumably Cucumber Ltd founded
that year as in nearly all other licenses in this project.

Including both Aslaks personal statement and that of Cucumber Ltd should
accurately reflect the copyright assignment.