[Core] Include stack traces in html report (#2862)
With cucumber/react-components#345 the html formatter started to
support rendering stack traces. Unfortunately this also broke the
regular rendering of stacktrace. And fixing this required including the
stack trace the xml report as well so that when the Convertor in messages
was fixed to always include the stacktrace, we wouldn't render the
stacktrace in the xml formatter twice.