Lazily start IncrementingUuidGenerator sessions (#2931)
Even when not used, the `IncrementingUuidGenerator` is instantiated with
each Cucumber execution. This somewhat fundamental to the way the
`ServiceLoader` mechanism works.

Each time an incrementing generator is created, a new session is started
for that generator. This means that after 255 executions all sessions
are exhausted.

Unfortunately, when using the JUnit Platform, Maven issues a discovery
request for each individual class. And Cucumber typically participates
in discovery along with JUnit Jupiter. So after 255 classes, Cucumber
fails discovery as seen in #2930.

Fixes #2930.