Include root cause when using DataTable.asList and friends
The `CucumberInvocationTargetException` was originally designed to be
used as a wrapper around `InvocationTargetException` to track the step
definition that caused the exception. The cause of
`InvocationTargetException` would then be rewritten to appear to have
originated from the step definition. This would effectively hide the
whole backend from view.

Unfortunately using DataTable.asList inside a step, this also invokes
the backend. And because `CucumberInvocationTargetException` did not
have cause, the cause of any exceptions would also be hidden.

By exposing the cause of the `InvocationTargetException` as the cause
of the `CucumberInvocationTargetException` and removing any frames
before the step definition was invoked we can clean this up somewhat.

```log
io.cucumber.datatable.CucumberDataTableException: 'java.util.List<io.cucumber.skeleton.StepDefinitions$Ingredient>' could not transform
| NAME  | QUANTITY | UNITS |
| Flour | 1        | KG    |
| Water | 0.65     | L     |
| Salt  | 0.08     | KG    |

	at io.cucumber.datatable.DataTableType.transform(DataTableType.java:158)
	at io.cucumber.datatable.DataTableTypeRegistryTableConverter.toListOrProblems(DataTableTypeRegistryTableConverter.java:158)
	at io.cucumber.datatable.DataTableTypeRegistryTableConverter.toList(DataTableTypeRegistryTableConverter.java:139)
	at io.cucumber.datatable.DataTable.asList(DataTable.java:199)
	at io.cucumber.skeleton.StepDefinitions.i_wait_hour(StepDefinitions.java:40)
	at ✽.I mix the following ingredients(classpath:io/cucumber/skeleton/belly.feature:6)
Caused by: io.cucumber.core.backend.CucumberInvocationTargetException
	at io.cucumber.java.Invoker.doInvoke(Invoker.java:73)
	at io.cucumber.java.Invoker.invoke(Invoker.java:24)
	at io.cucumber.java.AbstractGlueDefinition.invokeMethod(AbstractGlueDefinition.java:47)
	at io.cucumber.java.JavaDataTableTypeDefinition.lambda$createDataTableType$2(JavaDataTableTypeDefinition.java:47)
	at io.cucumber.datatable.DataTableType$TableEntryTransformerAdaptor.transform(DataTableType.java:342)
	at io.cucumber.datatable.DataTableType$TableEntryTransformerAdaptor.transform(DataTableType.java:319)
	at io.cucumber.datatable.DataTableType.transform(DataTableType.java:155)
	... 5 more
Caused by: java.lang.IllegalArgumentException: This message is never shown during test execution
	at io.cucumber.skeleton.StepDefinitions.mySystemEntry(StepDefinitions.java:29)
	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.base/java.lang.reflect.Method.invoke(Method.java:568)
	at io.cucumber.java.Invoker.doInvoke(Invoker.java:66)
	... 11 more
```

Unfortunately, it is not possible to remove the
`CucumberInvocationTargetException` entirely.

Fixes: #2948