[Core] Improve caching glue performance (#2971)
The `CachingGlue` performance has been improved by caching the following elements:

- parameter types
- data table types
- docString types
- step definitions

The cache is invalidated when a parameter type or a step definition is added/removed, or when the language changes.
The cache is not used when scenario-scoped glue is present (e.g. cucumber-java8).


This improves the `CachingGlue` performance.

On a personal project with about 250 step definitions and 1000 test scenarios, the performance is the following:

| Version | Test duration |
|---------|----------------|
|7.20.1 (original) | 8.2 seconds |
|7.21.0-SNAPSHOT (this PR) | 3.4 seconds (2.4x faster) |


---------

Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>
Co-authored-by: Julien Kronegg <julien@kronegg.ch>