Add custom UuidGenerator to Runtime.Builder (#3039)
Currently the UuidGenerator is looked up by SPI when creating the event
bus. This requires the UuidGenerator to be part of the classloader what
might not always be the case.

This now adds an option to set an actual instance for the UuidGenerator
and uses that in case it is specified.

Co-authored-by: Christoph Läubrich <christoph@laeubi-soft.de>
Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>