[JUnit Platform Engine] Warn when selected line does not exist (#3056)
It is possible to select a line in a classpath resource that does not
exist. For example:

```java
@Suite
@IncludeEngines("cucumber")
@SelectClasspathResource(value = "path/to/example.feature", line = 9001)
public class RunCucumberTest {
}
```

In a trivial suite this will result in an exception as no scenarios
could be found. In a non-trivial suite that does discover other
scenarios this error will go unnoticed.

By reporting this as a discovery issue and falling back to running the
whole feature we ensure that no-tests go unexecuted while also informing
the callers about this mis-configuration.