Warn when surefire naming strategy is used (#3067)
With Surefire 3.5.4 the heuristic for naming tests changed[1]. As a
result the `surefire` strategy no longer works as expected. But the
`long` strategy now does provide reasonable results.

On Surefire <= 3.5.2 with the `surefire` strategy the test results would
be rendered as:

```
<testcase name="a few cukes" classname="Belly" time="0.062">
```

This includes both the scenario and feature name. 


On Surefire >= 3.5.4 with the `surefire` strategy the test results will
be rendered as:

```
<testcase name="a few cukes" classname="RunCucumberTest" time="0.06">
```

This does not include the feature name. 


On Surefire >= 3.5.4 with the `long` strategy the test results will be
rendered as:

```
<testcase name="Belly - a few cukes" classname="RunCucumberTest" time="0.06">
```

This does the scenario and feature name and also the name of the Suite
that ran Cucumber. This is the optimal outcome.


1. https://github.com/apache/maven-surefire/pull/828

Closes: https://github.com/cucumber/cucumber-jvm/issues/2984