Generate annotations for Emoji dialect (#3062)
Step definition annotations and methods were not generated for the Emoji
dialect because Emoji are not valid Java identifiers. By making a small
exception in the script that generates these annotations we can instead
use their symbol names.

So when using an Emoji feature file:

```feature
# language: em
📚: 🙈🙉🙊

  📕: 💃
    😐🎸
```

This scenario can be made to pass with:

```java
public class StepDefinitions {
    @NeutralFace("\uD83C\uDFB8")
    public void guitar() {
    }
}
```

Closes: https://github.com/cucumber/cucumber-jvm/issues/3061

Additionally, by removing the `’` from the Texan keywords we can also
make the Texan dialect work. Like the `em` dialect `en-tx` is intended
 to be humorous. I hope all Texans will look at it with that spirit.