Prefer URIs with authority
There are two ways to obtain a URI of a file:

* `File.toURI()` -> `file:/path/to/file.feature`
* `Path.toUri()` -> `file:///path/to/file.feature`

While both URIs are equal, their string representation is not. This
poses some problems when comparing string based outputs.

Currently, when using the Cucumber JUnit Platform Engine the latter
variant is used. When using the CLI the former variant is used.

The choice between the two is fairly arbitrary. But the latter seems to
picked up by IDEA and turned into a clickable link. The former is not.